// src/app/models/references-data.ts
export interface ReferencesData {

  selectedRatingL: String   // Variable que guardará el valor seleccionado
  selectedRatingT: String   // Variable que guardará el valor seleccionado
  selectedRatingC: String   // Variable que guardará el valor seleccionado
  selectedRatingR: String   // Variable que guardará el valor seleccionado
  selectedRatingP: String   // Variable que guardará el valor seleccionado
  selectedRatingTE: String  // Variable que guardará el valor seleccionado
  selectedRatingCO: String  // Variable que guardará el valor seleccionado

  nombreReferencia: String,
  cargo: String,
  empresa: String,
  telefonosContacto: String

}